backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
tensorflow==2.16.1
numpy==1.26.4
opencv-python==4.9.0.80
python-dotenv==1.0.1

